requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
PyPDF2
pydantic
streamlit
